Tambah mode pada env

+++ controller/order_controller.go
@@ -7,6 +7,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"os"
 	"runtime"
 	"strconv"
 	"strings"
@@ -256,6 +257,7 @@ func (orderController *OrderController) ListDetail(param *utils.Param) map[strin
 }
 
 func (orderController *OrderController) UploadOrderProof(w http.ResponseWriter, r *http.Request) {
+	mode := os.Getenv("MODE")
 	id, err := GetInt64Param("id", w, r)
 	if err != nil {
 		utils.RespondError(w, utils.MessageErr(false, utils.ErrRequest, err.Error()), http.StatusBadRequest)
@@ -266,32 +268,36 @@ func (orderController *OrderController) UploadOrderProof(w http.ResponseWriter, 
 		utils.RespondError(w, utils.MessageErr(false, utils.ErrRequest, err.Error()), http.StatusBadRequest)
 		return
 	}
-	file, retval := utils.GetImageFile(w, r)
-	if retval.ErrCode != 0 {
-		utils.RespondError(w, utils.MessageErr(false, retval.ErrCode, retval.Message), http.StatusUnsupportedMediaType)
-		return
-	}
-	imageWritter := utils.GetImageWritter()
-	currentTime := time.Now()
 
-	retval = ViewModel(id, &orderController.Order)
-	if retval.ErrCode != 0 {
-		utils.Respond(w, utils.MessageErr(false, retval.ErrCode, retval.Message))
-		return
-	} else {
-		proofLink := orderController.Order.ProofLink
-		if proofLink != "" {
-			if err := imageWritter.DeleteFile(proofLink); err != nil {
-				utils.RespondError(w, utils.MessageErr(false, utils.ErrIO, err.Error()), http.StatusNotFound)
-				return
+	var fileName string
+	if mode != "DEV" {
+		file, retval := utils.GetImageFile(w, r)
+		if retval.ErrCode != 0 {
+			utils.RespondError(w, utils.MessageErr(false, retval.ErrCode, retval.Message), http.StatusUnsupportedMediaType)
+			return
+		}
+		imageWritter := utils.GetImageWritter()
+		currentTime := time.Now()
+
+		retval = ViewModel(id, &orderController.Order)
+		if retval.ErrCode != 0 {
+			utils.Respond(w, utils.MessageErr(false, retval.ErrCode, retval.Message))
+			return
+		} else {
+			proofLink := orderController.Order.ProofLink
+			if proofLink != "" {
+				if err := imageWritter.DeleteFile(proofLink); err != nil {
+					utils.RespondError(w, utils.MessageErr(false, utils.ErrIO, err.Error()), http.StatusNotFound)
+					return
+				}
 			}
 		}
-	}
 
-	fileName := "payment-" + strconv.Itoa(int(id)) + "-" + currentTime.Format("20060102150405")
-	if err := imageWritter.UploadFile(file, fileName); err != nil {
-		utils.RespondError(w, utils.MessageErr(false, utils.ErrIO, err.Error()), http.StatusUnsupportedMediaType)
-		return
+		fileName = "payment-" + strconv.Itoa(int(id)) + "-" + currentTime.Format("20060102150405")
+		if err := imageWritter.UploadFile(file, fileName); err != nil {
+			utils.RespondError(w, utils.MessageErr(false, utils.ErrIO, err.Error()), http.StatusUnsupportedMediaType)
+			return
+		}
 	}
 	resp := UpdateFieldMaster(id, orderController, func(m model.Model, db *gorm.DB) utils.StatusReturn {
 		order := m.(*model.Order)
